Rename the 'class' argument passed to add_settings_field() to 'instance'.
In WordPress 4.2, a new field was added to `add_settings_field()` with a key of 'class'. The purpose of the new argument was to make it possible to specify a CSS class for the setting container.

Unfortunately, the introduction of this new argument caused a conflict with a separate 'class' argument already in use. This PR renames that to 'instance' and all is right in the world.